pyproject.toml
[project]
name = "leo-cv-generator"
version = "1.0.0"
description = "Career CV and Cover Letter Generator from Knowledge Base"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Leo Zhang", email = "leozhang2056@gmail.com"}
]

dependencies = [
    "pyyaml==6.0.2",
    "playwright==1.58.0",
    "requests==2.32.3",
    "beautifulsoup4==4.14.3",
    "pypdf==6.7.5",
    "pydantic==2.12.5",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.1.0",
    "pytest-mock==3.15.1",
    "pygments==2.19.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
